+++ magda-gateway/views/admin.js
@@ -1,4 +1,5 @@
 const instanceURL = "/api/v0";
+let timeout = null;
 
 window.onload = function() {
     refresh();
@@ -18,13 +19,32 @@ async function refresh() {
             return;
         }
 
-        showMe(body.append("div"), me);
+        const sections = {
+            "My Information": showMe.bind(null, me),
+            "Header Logo": showHeaderLogo,
+            "Header Navigation": showHeaderNavigation,
+            "Header Taglines": showHeaderTaglines,
+            "Homepage Highlights": showHomeHighlights,
+            "Homepage Stories": showHomeStories,
+            "CSV Data": showSpreadsheets,
+            Connectors: showConnectors,
+            Language: showLanguage,
+            "Footer Navigation": showFooterNavigation,
+            "Footer Copyright": showFooterCopyright
+        };
 
-        showConfig(body.append("div"));
-
-        showSpreadsheets(body.append("div"));
-
-        showConnectors(body.append("div"));
+        const section = body.append("select");
+        const sectionBody = body.append("div");
+        for (let key of Object.keys(sections)) {
+            section.append("option").text(key);
+        }
+        section.on("change", () => {
+            if (timeout) {
+                clearTimeout(timeout);
+            }
+            sections[section.property("value")](sectionBody.text("Loading..."));
+        });
+        section.on("change")();
     } catch (e) {
         body.append("pre").text(e);
 
@@ -39,8 +59,14 @@ async function refresh() {
     }
 }
 
-function showConfig(body) {
-    body.append("h2").text("Configuration");
+function showMe(me, body) {
+    body.text("");
+    body.append("h2").text("My Information");
+    body.append("pre").text(JSON.stringify(me, null, 2));
+}
+
+function showHeaderLogo(body) {
+    body.append("h2").text("Header Logo");
 
     const table = body.append("table");
 
@@ -57,105 +83,282 @@ function showConfig(body) {
     row = table.append("tr");
     row.append("td").text("Mobile Logo");
     imageConfig(row.append("td"), "header/logo-mobile");
-
-    showHeaderNavigation(body.append("div"));
 }
 
-async function showHeaderNavigation(body) {
-    body.text("Loading...");
+function showHeaderNavigation(body) {
+    showJsonEditor(body, {
+        label: "Header Navigation",
+        idPattern: "header/navigation/*",
+        schema: headerNavigationSchema,
+        allowDelete: true,
+        allowAdd: true,
+        newId: () => `header/navigation/${Date.now()}`
+    });
+}
 
-    let files = await request("GET", "/api/v0/content/all?inline=true");
-    files = files
-        .filter(x => x.id.match(/^header\/navigation\//))
-        .sort((a, b) => a.content.order - b.content.order);
+function showHeaderTaglines(body) {
+    showJsonEditor(body, {
+        label: "Header Taglines",
+        idPattern: "home/tagline/*",
+        schema: headerTaglineSchema,
+        allowDelete: false,
+        allowAdd: false
+    });
+}
+
+function showHomeStories(body) {
+    showJsonEditor(body, {
+        label: "Home Stories",
+        idPattern: "home/stories/*",
+        schema: homeStorySchema,
+        allowDelete: true,
+        allowAdd: true,
+        newId: () => `home/stories/${Date.now()}`,
+        extraControls: (parent, file) => {
+            parent.append("h4").text("Story Image");
+            imageConfig(
+                parent.append("div"),
+                file.id.replace(/^home\/stories/, "home/story-images")
+            );
+        }
+    });
+}
+
+function showHomeHighlights(body) {
+    showJsonEditor(body, {
+        label: "Home Highlights",
+        idPattern: "home/highlights/*",
+        schema: homeHighlightSchema,
+        allowDelete: true,
+        allowAdd: true,
+        newId: () => `home/highlights/${Date.now()}`,
+        extraControls: (parent, file) => {
+            parent.append("h4").text("Highlight Images width > 0");
+            imageConfig(
+                parent.append("div"),
+                file.id.replace(/^home\/highlights/, "home/highlight-images") +
+                    "/0w"
+            );
+            parent.append("h4").text("Highlight Images width > 720");
+            imageConfig(
+                parent.append("div"),
+                file.id.replace(/^home\/highlights/, "home/highlight-images") +
+                    "/720w"
+            );
+            parent.append("h4").text("Highlight Images width > 1080");
+            imageConfig(
+                parent.append("div"),
+                file.id.replace(/^home\/highlights/, "home/highlight-images") +
+                    "/1080w"
+            );
+            parent.append("h4").text("Highlight Images width > 1440");
+            imageConfig(
+                parent.append("div"),
+                file.id.replace(/^home\/highlights/, "home/highlight-images") +
+                    "/1440w"
+            );
+            parent.append("h4").text("Highlight Images width > 2160");
+            imageConfig(
+                parent.append("div"),
+                file.id.replace(/^home\/highlights/, "home/highlight-images") +
+                    "/2160w"
+            );
+            parent.append("h4").text("Highlight Images width > 2880");
+            imageConfig(
+                parent.append("div"),
+                file.id.replace(/^home\/highlights/, "home/highlight-images") +
+                    "/2880w"
+            );
+        }
+    });
+}
+
+function showLanguage(body) {
+    showJsonEditor(body, {
+        label: "Language Items",
+        idPattern: "lang/en/*",
+        schema: languageSchema,
+        allowDelete: true,
+        allowAdd: true,
+        allowIdFieldInput: true,
+        newId: id => `lang/en/${id}`
+    });
+}
+
+function showFooterNavigation(body) {
+    body.text("");
+    showJsonEditor(body.append("div"), {
+        label: "Medium Footer Navigation Categories",
+        idPattern: "footer/navigation/medium/category/*",
+        schema: footerCategory,
+        allowDelete: true,
+        allowAdd: true,
+        newId: id => `footer/navigation/medium/category/${Date.now()}`,
+        extraControls: (parent, file) => {
+            const prefix = file.id.substr(
+                "footer/navigation/medium/category/".length
+            );
+            showJsonEditor(parent.append("div").style("padding-left", "5em"), {
+                label: "Category Menu Items",
+                idPattern: `footer/navigation/medium/category-links/${prefix}/*`,
+                schema: footerLink,
+                allowDelete: true,
+                allowAdd: true,
+                newId: id =>
+                    `footer/navigation/medium/category-links/${prefix}/${Date.now()}`
+            });
+        }
+    });
+    showJsonEditor(body.append("div"), {
+        label: "Small Footer Navigation Categories",
+        idPattern: "footer/navigation/small/category/*",
+        schema: footerCategory,
+        allowDelete: true,
+        allowAdd: true,
+        newId: id => `footer/navigation/small/category/${Date.now()}`,
+        extraControls: (parent, file) => {
+            const prefix = file.id.substr(
+                "footer/navigation/small/category/".length
+            );
+            showJsonEditor(parent.append("div").style("padding-left", "5em"), {
+                label: "Category Menu Items",
+                idPattern: `footer/navigation/small/category-links/${prefix}/*`,
+                schema: footerLink,
+                allowDelete: true,
+                allowAdd: true,
+                newId: id =>
+                    `footer/navigation/small/category-links/${prefix}/${Date.now()}`
+            });
+        }
+    });
+}
+
+function showFooterCopyright(body) {
+    showJsonEditor(body, {
+        label: "Footer Copyright",
+        idPattern: "footer/copyright/*",
+        schema: footerCopyright,
+        allowDelete: true,
+        allowAdd: true,
+        newId: id => `footer/copyright/${Date.now()}`
+    });
+}
+
+async function showJsonEditor(body, options) {
+    let files = await request(
+        "GET",
+        `/api/v0/content/all?id=${options.idPattern}&inline=true`
+    );
+    if (files.length && files[0].order !== undefined) {
+        files = files.sort((a, b) => a.content.order - b.content.order);
+    }
 
     body.text("");
 
-    body.append("h2").text("Header Navigation");
+    body.append("h2").text(options.label);
 
     if (files.length > 0) {
-        files.forEach(file => {
+        files.forEach((file, index, files) => {
             const container = jsoneditor(
                 body.append("div").style("border", "10px solid black"),
-                "Edit",
-                headerNavigationSchema,
-                file.content,
-                "Save",
-                async newObj => {
-                    console.log(
-                        "DONE",
-                        name,
-                        await request(
-                            "POST",
-                            `/api/v0/content/${file.id}`,
-                            newObj,
-                            "application/json"
-                        )
-                    );
-                    showHeaderNavigation(body);
+                {
+                    title: file.id,
+                    schema: options.schema,
+                    startval: file.content,
+                    label: "Save",
+                    callback: async newObj => {
+                        console.log(
+                            "DONE",
+                            name,
+                            await request(
+                                "POST",
+                                `/api/v0/content/${file.id}`,
+                                newObj,
+                                "application/json"
+                            )
+                        );
+                        showJsonEditor(body, options);
+                    }
                 }
             );
-            container
-                .append("button")
-                .text("Delete")
-                .on("click", async () => {
-                    console.log(
-                        "DONE",
-                        name,
-                        await request("DELETE", `/api/v0/content/${file.id}`)
-                    );
-                    showHeaderNavigation(body);
-                });
+            if (options.allowDelete) {
+                container
+                    .append("button")
+                    .text("Delete")
+                    .on("click", async () => {
+                        console.log(
+                            "DONE",
+                            name,
+                            await request(
+                                "DELETE",
+                                `/api/v0/content/${file.id}`
+                            )
+                        );
+                        showJsonEditor(body, options);
+                    });
+            }
+            if (options.extraControls) {
+                options.extraControls(container, file, index, files);
+            }
         });
     } else {
         body.append("P").text("You got none!");
     }
 
-    jsoneditor(
-        body.append("div"),
-        "Add new",
-        headerNavigationSchema,
-        null,
-        "Add",
-        async newObj => {
-            console.log(
-                "DONE",
-                name,
-                await request(
-                    "POST",
-                    `/api/v0/content/header/navigation/${Date.now()}`,
-                    newObj,
-                    "application/json"
-                )
-            );
-            showHeaderNavigation(body);
+    if (options.allowAdd) {
+        let idField;
+        if (options.allowIdFieldInput) {
+            const container = body.append("p");
+            container.append("strong").text("ID: ");
+            idField = container.append("input");
         }
-    );
+        jsoneditor(body.append("div"), {
+            title: "Add new",
+            schema: options.schema,
+            startval: null,
+            label: "Add",
+            callback: async newObj => {
+                console.log(
+                    "DONE",
+                    name,
+                    await request(
+                        "POST",
+                        `/api/v0/content/${options.newId(
+                            idField ? idField.property("value") : undefined
+                        )}`,
+                        newObj,
+                        "application/json"
+                    )
+                );
+                showJsonEditor(body, options);
+            }
+        });
+    }
 }
 
-function jsoneditor(parent, title, schema, startval, label, callback) {
+function jsoneditor(parent, options) {
     var element = parent.node();
     var editor = new JSONEditor(element, {
-        schema: Object.assign({}, schema, {
-            title,
+        schema: Object.assign({}, options.schema, {
+            title: options.title,
             options: {
                 remove_empty_properties: true
             }
         }),
-        startval,
+        startval: options.startval,
         disable_edit_json: true,
         disable_collapse: true,
         disable_properties: false,
         show_errors: "always",
-        form_name_root: title,
+        form_name_root: options.title,
         keep_oneof_values: false,
         no_additional_properties: true
     });
-    var button = parent.append("button").text(label);
+    var button = parent.append("button").text(options.label);
     button.on("click", async () => {
         const value = await validate();
         if (value) {
-            callback(value);
+            options.callback(value);
         }
     });
     async function validate() {
@@ -174,14 +377,12 @@ function jsoneditor(parent, title, schema, startval, label, callback) {
     return parent;
 }
 
-function showMe(body, me) {
-    body.append("h2").text("User");
-    body.append("pre").text(JSON.stringify(me, null, 2));
-}
-
 function imageConfig(body, name) {
+    body.text("");
+
     body.append("img")
         .attr("src", `${instanceURL}/content/${name}.bin`)
+        .attr("onerror", `this.alt='NONE'`)
         .style("max-height", `70px`)
         .style("max-width", `367px`);
 
@@ -203,7 +404,7 @@ function imageConfig(body, name) {
                         data,
                         file.type
                     );
-                    window.location = window.location;
+                    imageConfig(body, name);
                 };
                 fileReader.readAsArrayBuffer(file);
             };
@@ -312,6 +513,8 @@ async function showSpreadsheets(body) {
     body.append("H3").text("Upload new");
 
     spreadsheetConfig(body);
+
+    showConnectors(body.append("div"));
 }
 
 async function deleteContent(name) {
@@ -347,6 +550,7 @@ async function createConnector(name) {
 }
 
 function showConnectors(body) {
+    body.text("");
     body.text("Loading...");
 
     async function refresh() {
@@ -407,7 +611,7 @@ function showConnectors(body) {
 
         // body.append("pre").text(JSON.stringify(connectors, null, 2));
 
-        setTimeout(refresh, running ? 1000 : 5000);
+        timeout = setTimeout(refresh, running ? 1000 : 5000);
     }
 
     refresh();
@@ -499,4 +703,115 @@ const headerNavigationSchema = {
         }
     ],
     required: ["order"]
+};
+
+const headerTaglineSchema = {
+    type: "string"
+};
+
+const languageSchema = {
+    type: "string"
+};
+
+const homeStorySchema = {
+    type: "object",
+    properties: {
+        title: {
+            type: "string",
+            minLength: 1
+        },
+        titleUrl: {
+            type: "string",
+            minLength: 1
+        },
+        order: {
+            type: "number"
+        },
+        content: {
+            type: "string",
+            minLength: 1,
+            format: "markdown"
+        }
+    },
+    required: ["title", "order", "content"]
+};
+
+const homeHighlightSchema = {
+    type: "object",
+    properties: {
+        text: {
+            type: "string",
+            minLength: 1
+        },
+        url: {
+            type: "string",
+            minLength: 1
+        }
+    },
+    required: ["text", "url"]
+};
+
+const footerCategory = {
+    type: "object",
+    properties: {
+        order: {
+            type: "number",
+            default: 1
+        },
+        label: {
+            type: "string",
+            minLength: 1
+        }
+    },
+    required: ["order", "label"]
+};
+
+const footerLink = {
+    type: "object",
+    properties: {
+        order: {
+            type: "number",
+            default: 1
+        },
+        label: {
+            type: "string",
+            minLength: 1
+        },
+        href: {
+            type: "string",
+            minLength: 1
+        }
+    },
+    required: ["order", "label", "href"]
+};
+
+const footerCopyright = {
+    type: "object",
+    properties: {
+        order: {
+            type: "number",
+            default: 1
+        },
+        href: {
+            type: "string",
+            minLength: 1
+        },
+        logoSrc: {
+            type: "string",
+            minLength: 1
+        },
+        logoClassName: {
+            type: "string",
+            default: ""
+        },
+        logoAlt: {
+            type: "string",
+            default: ""
+        },
+        htmlContent: {
+            type: "string",
+            minLength: 1
+        }
+    },
+    required: ["order", "href", "logoSrc", "htmlContent"]
 };