+++ magda-web-client/src/AppContainer.js
@@ -8,6 +8,7 @@ import Banner from "./UI/Banner";
 import Footer from "./Components/Footer/Footer";
 
 import { requestWhoAmI } from "./actions/userManagementActions";
+import { fetchContent } from "./actions/contentActions";
 import Notification from "./UI/Notification";
 import { hideTopNotification } from "./actions/topNotificationAction";
 
@@ -23,12 +24,12 @@ import "./AppContainer.css";
 class AppContainer extends React.Component {
     componentDidMount() {
         this.props.requestWhoAmI();
+        this.props.fetchContent();
     }
 
     render() {
-        const footerNavs = config.footerNavigation;
         return (
-            <ReactDocumentTitle title={config.appName}>
+            <ReactDocumentTitle title={this.props.strings.applicationName}>
                 <div className="au-grid wrapper">
                     <div>
                         <AUskipLink
@@ -55,7 +56,7 @@ class AppContainer extends React.Component {
                         </Switch>
                     </div>
 
-                    <Footer footerNavs={footerNavs} />
+                    <Footer />
 
                     {this.props.topNotification.visible ? (
                         <Notification
@@ -85,7 +86,8 @@ class AppContainer extends React.Component {
 
 const mapStateToProps = state => {
     return {
-        topNotification: state.topNotification
+        topNotification: state.topNotification,
+        strings: state.content.strings
     };
 };
 
@@ -93,6 +95,7 @@ const mapDispatchToProps = dispatch => {
     return bindActionCreators(
         {
             requestWhoAmI: requestWhoAmI,
+            fetchContent,
             hideTopNotification
         },
         dispatch