+++ magda-web-client/src/Pages/HomePageComponents/StoryBox.js
@@ -1,27 +1,8 @@
 import React, { Component } from "react";
-import PropTypes from "prop-types";
-import { connect } from "react-redux";
-import { bindActionCreators } from "redux";
-import { fetchHomepageStory } from "../../actions/homePageStoriesActions";
-import { safeLoadFront } from "yaml-front-matter/dist/yamlFront";
 import MarkdownViewer from "../../UI/MarkdownViewer";
-import { config } from "../../config";
 import "./StoryBox.css";
 
-const baseUrl = config.homePageConfig.baseUrl
-    ? config.homePageConfig.baseUrl
-    : "";
-
 class StoryBox extends Component {
-    constructor(props) {
-        super(props);
-        this.state = {};
-    }
-
-    componentDidMount() {
-        this.props.fetchHomepageStory(this.props.idx);
-    }
-
     getClickableElement(el, url) {
         if (!url) return el;
         return (
@@ -37,18 +18,17 @@ class StoryBox extends Component {
     }
 
     renderStoryboxBody() {
-        const info = this.props.stories[this.props.idx];
-        if (info.isFetching) return <div>Loading...</div>;
-        if (info.isError) return <div>{info.error.message}</div>;
-        if (!info.content) return <div>No content available.</div>;
-        const content = safeLoadFront(info.content);
+        const { story } = this.props;
+        if (!story) return null;
+        const { content, image } = story;
+        if (!content) return null;
         const innerBody = (
             <div className="story-box-body">
-                {content.titleImage
+                {image
                     ? this.getClickableElement(
                           <img
                               className="story-title-image"
-                              src={`${baseUrl}${content.titleImage}`}
+                              src={`${image}`}
                               alt="title"
                           />,
                           content.titleUrl
@@ -61,7 +41,7 @@ class StoryBox extends Component {
                               content.titleUrl
                           )
                         : null}
-                    <MarkdownViewer markdown={content.__content} />
+                    <MarkdownViewer markdown={content.content} />
                 </div>
             </div>
         );
@@ -84,26 +64,4 @@ class StoryBox extends Component {
     }
 }
 
-StoryBox.propTypes = {
-    idx: PropTypes.number.isRequired
-};
-
-function mapStateToProps(state) {
-    return {
-        stories: state.homepageStories
-    };
-}
-
-const mapDispatchToProps = dispatch => {
-    return bindActionCreators(
-        {
-            fetchHomepageStory: fetchHomepageStory
-        },
-        dispatch
-    );
-};
-
-export default connect(
-    mapStateToProps,
-    mapDispatchToProps
-)(StoryBox);
+export default StoryBox;