+++ magda-web-client/src/helpers/record.js
@@ -356,9 +356,10 @@ export function parseDataset(dataset?: RawDataset): ParsedDataset {
         ? aspects["dataset-publisher"]["publisher"]
         : defaultPublisher;
     const contactPoint: string = aspects["dcat-dataset-strings"].contactPoint;
-
     const source: string = aspects["source"]
-        ? aspects["source"]["name"]
+        ? aspects["source"]["type"] !== "csv-dataset"
+            ? aspects["source"]["name"]
+            : undefined
         : defaultDatasetAspects["source"]["name"];
 
     function calcQuality(qualityAspect) {
@@ -371,6 +372,9 @@ export function parseDataset(dataset?: RawDataset): ParsedDataset {
     const linkedDataRating: number = aspects["dataset-quality-rating"]
         ? calcQuality(aspects["dataset-quality-rating"])
         : 0;
+    const hasQuality: boolean = aspects["dataset-quality-rating"]
+        ? true
+        : false;
 
     const distributions = distribution["distributions"].map(d => {
         const distributionAspects = Object.assign(
@@ -436,6 +440,7 @@ export function parseDataset(dataset?: RawDataset): ParsedDataset {
         temporalCoverage,
         publisher,
         error,
-        linkedDataRating
+        linkedDataRating,
+        hasQuality
     };
 }